Split feed cleaner and job
